ChafaFrame, ChafaImage: Share innards with rest of library
